Refactor to split settings module and avoid circular references
