+++ esg_fastapi/__init__.py
@@ -9,9 +9,6 @@ The stack consists of a FastAPI application that is configured to serve multiple
 The main API also includes an observability endpoint at the "/observability" sub-route. This endpoint is used to expose OpenTelemetry generated by the API, such as health checks, request counts, response times, and error rates.
 """
 
-# NOTE: This needs to happen as early as possible so that any logging that happens
-# during initialization is instrumented. We do it here instead of `__main__.py`
-# since it has to import this module, thus this happens first
-from opentelemetry.instrumentation.logging import LoggingInstrumentor
+from .configuration import settings
 
-LoggingInstrumentor().instrument()  # noqa
+__all__ = ["settings"]
