+++ esg_fastapi/__main__.py
@@ -5,10 +5,14 @@ from typing import Self
 from fastapi import FastAPI
 from gunicorn.app.base import BaseApplication
 from gunicorn.arbiter import Arbiter
+from opentelemetry import trace
+from opentelemetry.sdk.resources import Resource
+from opentelemetry.sdk.trace import TracerProvider
+from opentelemetry.sdk.trace.export import BatchSpanProcessor, ConsoleSpanExporter
 
 from esg_fastapi import settings
 
-from . import api
+from .api.main import app_factory
 
 
 class Server(BaseApplication):
@@ -31,7 +35,20 @@ class Server(BaseApplication):
         We use a factory function instead of importing an instantiated app so that Gunicorn can reload
         the app on the fly as needed.
         """
-        return api.wsgi_factory()
+        return app_factory()
 
 
-Arbiter(Server()).run()
+# TODO: more accurate init: https://opentelemetry-python.readthedocs.io/en/latest/api/trace.html#opentelemetry.trace.TracerProvider
+
+tracer: trace.Tracer = trace.get_tracer(
+    instrumenting_module_name="esg-fastapi",
+    tracer_provider=TracerProvider(
+        resource=Resource.create(),
+        active_span_processor=BatchSpanProcessor(
+            span_exporter=ConsoleSpanExporter()  # type: ignore -- library typing is overly specific
+        ),
+    ),
+)
+
+with tracer.start_as_current_span("main") as span:
+    Arbiter(Server()).run()
