+++ esg_fastapi/api/versions/__init__.py
@@ -5,10 +5,10 @@ This module provides a mechanism for discovering and importing all versioned API
 The discovered submodules are then added to the main application using the `app.mount()` method. This allows developers to easily extend and customize the functionality of the main application by adding new submodules that provide additional API endpoints.
 """
 
-import pkgutil
-from importlib import import_module
+# import pkgutil
+# from importlib import import_module
 
-discovered = []
-for submodule in pkgutil.iter_modules(__path__):
-    if submodule.name.lstrip("v").isdigit():
-        discovered.append(import_module("." + submodule.name, __package__))
+# discovered = []
+# for submodule in pkgutil.iter_modules(__path__):
+#     if submodule.name.lstrip("v").isdigit():
+#         discovered.append(import_module("." + submodule.name, __package__))
