+++ esg_fastapi/api/versions/v1/__init__.py
@@ -3,6 +3,6 @@
 Future versions will extend and inherit from this class.
 """
 
-from .routes import app
+# from .routes import app_factory
 
-__all__ = ["app"]
+# __all__ = ["app_factory"]
