+++ tests/test_esgsearch_parity.py
@@ -9,12 +9,7 @@ from pytest_bdd import given, scenarios, then, when
 from pytest_bdd.parsers import parse
 from pytest_mock import MockerFixture
 
-from esg_fastapi import api
-from esg_fastapi.api.versions.v1.models import (
-    ESGSearchResponse,
-    GlobusSearchQuery,
-    GlobusSearchResult,
-)
+from esg_fastapi.api.versions.v1.models import ESGSearchResponse, GlobusSearchQuery, GlobusSearchResult
 
 scenarios("ESGSearch_Parity")
 
@@ -45,7 +40,9 @@ def load_example(query_example: Path) -> SearchParityFixture:
 @when("the request is sent to ESG FastAPI", target_fixture="responses")
 def send_request(json_example: SearchParityFixture, mocker: MockerFixture) -> ComparisonFixture:
     """Send request to ESG FastAPI and add its response to the fixture."""
-    client = TestClient(api.wsgi_factory())
+    from esg_fastapi.api.main import app_factory
+
+    client = TestClient(app_factory())
     mocker.patch(
         "esg_fastapi.api.versions.v1.routes.SearchClient.post_search",
         return_value=mocker.Mock(data=json_example["globus_response"]),
